requirements.txt
uuid
pathlib
Flask==2.0.3
numpy==1.21.6
Pillow==9.0.1
transformers==4.18.0
pymongo==4.1.0
scipy==1.7.3
tensorflow-intel==2.10.0
tensorboard==2.10.0
protobuf==3.19.6
scikit-learn==1.0.2 
Werkzeug==2.2.2